requirements.txt
# install all with: sudo pip install -r requirements.txt --upgrade
# or by calling: pip install --user -r requirements.txt
soco
spi
pyserial
nfcpy
Flask
flask-bootstrap
flask-nav